requirements.txt
Django==1.8.2
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.20
django-celery==3.1.16
djangorestframework==3.1.3
kombu==3.0.26
psycopg2==2.6.1
pytz==2015.4
wsgiref==0.1.2
